rename queries.ts -> run-queries.ts
